shio: Backport #6248: Auto-encode audio track for Whisper
Whisper.cpp requires a very specific input audio format. Instead of
relying on adopters working around this in the workflow and encoding
tracks in advance, this patch includes the encoding in the operation
itself.

This not only has the advantage that the configuration is much easier,
but it also means that only the tracks selected to actually be used for
generating the subtitles are encoded.