shio: Backport #6499: Use system user for repopulate search index
The previous fix from #6490 did not fully address the issue, as the
`getSeries` service is dependent on the current organization context
from the `securityService`. Consequently, related series were not
retrievable.

To resolve this, a switch to the corresponding system user
was implemented. This change ensures that the correct organization
context is used, thereby granting all necessary permissions to reindex
both series and events.