shio: Backport #6885: Fix snapshot deletion for remote only MediaPackages
Deletion of snapshots will currently fail if all snapshots live in
remote asset stores. When trying to delete from the local asset store
the base directory of the MediaPackage will not be found leading to an
exception when trying to concatenate a path. This patch adds a null
check that will skip the deletion from the local asset store if the base
directory cannot be found.